requirements.txt
python-dotenv
tweepy
pyyaml
